Fix zooming issue.
Here I allow the zoom if the user wanted to zoom in, as always the user
can zoom in as long as the map is already displayed on the screen.

When the map is bounded to certain coordinates, some time the map
cannot be zoomed in if the zooming is taking place beside the boarders.

Note that when the zoom is at its max, the zoom will take place inside
the tile image itself.